edited the view budgets command to display the final budget as a floating point number with 2 sf (since it is a monetary value)

+++ src/main/java/seedu/address/storage/XmlAdaptedFinalClubBudget.java
@@ -41,7 +41,7 @@ public class XmlAdaptedFinalClubBudget {
      */
     public XmlAdaptedFinalClubBudget(FinalClubBudget source) {
         clubName = source.getClubName().toString();
-        allocatedBudget = Integer.toString(source.getAllocatedBudget());
+        allocatedBudget = Double.toString(source.getAllocatedBudget());
     }
 
     /**
